working version of users tracking but without on screen orientation chan...

+++ PhotoShoterModule/src/main/java/com/photoshoter/SocketClient.java
@@ -116,8 +116,8 @@ public class SocketClient {
             JSONObject positionJson = json.getJSONObject("position");
 
 
-            double latitude = positionJson.getInt("lat");
-            double longitude = positionJson.getInt("long");
+            double latitude = positionJson.getDouble("lat");
+            double longitude = positionJson.getDouble("long");
             Location loc = new Location("socketio");
             loc.setLatitude(latitude);
             loc.setLongitude(longitude);
